feat: add landing page
Previously, the header would contain links to »Notizen«, »Danksagung«
and »Dashboard«. We now reduced that to »Leitfaden« only, and will put
the other links into the footer.

We also introduced a new color `surface`. The previous `tertiary` color
was too dark to be used as a »theme« color. That previous `tertiary`
color is now the `surface` color.